Make ComposeViewport API common between k/js and k/wasm
Previously, they were not common - it was not possible to call this API from a shared webMain-like source set.

One of the overloads has no w3c types, and therefore it can be web-common for free.

Also, updated our demo: fun main is common now.
