+++ compose/mpp/demo/src/wasmJsMain/kotlin/androidx/compose/mpp/demo/Utils.wasm.kt
@@ -19,6 +19,15 @@ package androidx.compose.mpp.demo
 
 import androidx.compose.foundation.internal.readText
 import androidx.compose.ui.platform.ClipEntry
+import kotlin.wasm.unsafe.UnsafeWasmMemoryApi
+import kotlin.wasm.unsafe.withScopedMemoryAllocator
+import kotlinx.browser.window
+import kotlinx.coroutines.Deferred
+import kotlinx.coroutines.asDeferred
+import kotlinx.coroutines.await
+import org.khronos.webgl.ArrayBuffer
+import org.khronos.webgl.Int8Array
+import org.w3c.fetch.Response
 
 actual suspend fun ClipEntry?.getPlainText(): String? {
     return this?.readText()
@@ -26,4 +35,35 @@ actual suspend fun ClipEntry?.getPlainText(): String? {
 
 actual fun createClipEntryWithPlainText(text: String): ClipEntry {
     return ClipEntry.withPlainText(text)
+}
+
+private suspend fun loadResAsync(url: String): Deferred<ArrayBuffer> {
+    return window.fetch(url).await<Response>().arrayBuffer().asDeferred()
+}
+
+suspend fun loadRes(url: String): ArrayBuffer {
+    return loadResAsync(url).await()
+}
+
+fun ArrayBuffer.toByteArray(): ByteArray {
+    val source = Int8Array(this, 0, byteLength)
+    return jsInt8ArrayToKotlinByteArray(source)
+}
+
+private fun wasmExportsMemoryBuffer(): ArrayBuffer = js("wasmExports.memory.buffer")
+private fun jsExportInt8ArrayToWasm(destination: ArrayBuffer, src: Int8Array, size: Int, dstAddr: Int) {
+    val mem8 = Int8Array(destination, dstAddr, size)
+    mem8.set(src)
+}
+
+internal fun jsInt8ArrayToKotlinByteArray(x: Int8Array): ByteArray {
+    val size = x.length
+
+    @OptIn(UnsafeWasmMemoryApi::class)
+    return withScopedMemoryAllocator { allocator ->
+        val memBuffer = allocator.allocate(size)
+        val dstAddress = memBuffer.address.toInt()
+        jsExportInt8ArrayToWasm(wasmExportsMemoryBuffer(),  x, size, dstAddress)
+        ByteArray(size) { i -> (memBuffer + i).loadByte() }
+    }
 }
