+++ compose/mpp/demo/src/webMain/resources/NotoColorEmoji.ttf
