+++ compose/mpp/demo/src/webMain/resources/NotoSansSC-Regular.ttf
